nothing new here

+++ src/Models/EasyEquationGenerator.java
@@ -13,7 +13,7 @@ public class EasyEquationGenerator implements EquationGenerator {
 	
 	@Override
 	public Equation generateEquation() {
-		int initialSpeed = (int) MiscUtils.rand(100, 200);
+		int initialSpeed = (int) ((int) MiscUtils.rand(100, 200));
 		int startPoint = (int) MiscUtils.rand(0, screenWidth / 2);
 		double maxAngle = Math.asin(Math.sqrt((2 * 9.8 * screenHeight) / (initialSpeed * initialSpeed)));
 		double projectionAngle = MiscUtils.rand(Math.PI / 9, maxAngle);
